+++ js/scripts.js
@@ -1,5 +1,6 @@
 // import im from 'include-media-export/include-media.js';
 // import fitvids from 'fitvids/dist/fitvids.js';
+import 'instant.page';
 
 ( function() {
 	// Remove the 'no-js' <body> class